codebert/requirements.txt
# Core ML stack (RTX 5090 / Blackwell sm_120 needs CUDA 12.8+)
# On RunPod, pick nvidia/cuda:12.8.0-cudnn-devel-ubuntu22.04 base image.
# Install torch with the cu128 index:
#   pip install --index-url https://download.pytorch.org/whl/cu128 torch==2.6.0
torch>=2.6.0
transformers>=4.45.0
accelerate>=1.0.0
datasets>=2.20.0
safetensors>=0.4.5

# Tokenization / parsing
tree_sitter==0.21.3
tree_sitter_languages==1.10.2
markdown-it-py>=3.0.0

# Data
pandas>=2.2.0
pyarrow>=16.0.0
numpy>=1.26.0
datasketch>=1.6.5
scikit-learn>=1.4.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
click>=8.1.0

# Logging (optional; code falls back to JSONL/TensorBoard if wandb missing)
wandb>=0.17.0; python_version >= "3.9"
tensorboard>=2.15.0

# Plotting + HP search
matplotlib>=3.8.0
optuna>=3.6.0
